requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
python-multipart
